requirements.txt
asgiref==3.7.2
Django==4.2.7
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-cuser==2017.3.16
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-tinymce==3.6.1
modeltranslation==0.25
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
